Permission management

+++ jbpm-gwt/jbpm-gwt-form-builder/src/main/java/org/jbpm/formbuilder/server/RESTMenuService.java
@@ -145,27 +145,31 @@ public class RESTMenuService extends RESTBaseService {
     }
     
     @DELETE @Path("/items")
-    public Response deleteMenuItem(SaveMenuItemDTO dto) {
+    public Response deleteMenuItem(SaveMenuItemDTO dto, @Context HttpServletRequest request) {
         init();
         try {
-            MenuItemDescription menuItem = toMenuItemDescription(dto, false);
-            Map<String, List<MenuItemDescription>> items = menuService.listMenuItems();
-            List<MenuItemDescription> group = items.get(dto.getGroupName());
-            if (group == null || group.isEmpty()) {
-                return Response.noContent().build();
-            }
-            boolean found = false;
-            for (MenuItemDescription desc : group) {
-                if (desc.getName().equals(dto.getName())) {
-                    found = true;
-                    break;
-                }
-            }
-            if (!found) {
-                return Response.status(Status.CONFLICT).build();
-            }
-            menuService.deleteMenuItem(dto.getGroupName(), menuItem);
-            return Response.status(Status.ACCEPTED).build();
+        	if (RESTUserService.hasDesignerPrivileges(request)) {
+        		MenuItemDescription menuItem = toMenuItemDescription(dto, false);
+        		Map<String, List<MenuItemDescription>> items = menuService.listMenuItems();
+        		List<MenuItemDescription> group = items.get(dto.getGroupName());
+        		if (group == null || group.isEmpty()) {
+        			return Response.noContent().build();
+        		}
+        		boolean found = false;
+        		for (MenuItemDescription desc : group) {
+        			if (desc.getName().equals(dto.getName())) {
+        				found = true;
+        				break;
+        			}
+        		}
+        		if (!found) {
+        			return Response.status(Status.CONFLICT).build();
+        		}
+        		menuService.deleteMenuItem(dto.getGroupName(), menuItem);
+        		return Response.status(Status.ACCEPTED).build();
+        	} else {
+        		return Response.status(Status.UNAUTHORIZED).build();
+        	}
         } catch (MenuServiceException e) {
             return error("Couldn't delete menu item " + dto.getGroupName() + ":" + dto.getName(), e);
         }
